Change the file processing behaviour
Instead of applying regular expressions to the entire file content to extract
values, the file content is now split up into lines.

Each line is processed separately using the field patterns. If multiple lines
are matched by a pattern, the additional lines are omitted and a warning is
printed.

A special case is the first line, which is directly used as password if
`FIRSTLINE_IS_LOGIN_PASSWORD` is set to `True`.

All lines not matching any pattern are put into a configurable column
`FALLBACK_FIELD`, which is set to `notes` by default.